Fix fast file system and pfs3 disk read and write above 4gb.

+++ src/Hst.Amiga/FileSystems/FastFileSystem/FastFileSystemFormatter.cs
@@ -70,9 +70,8 @@
             
             BlockHelper.UpdateBitmaps(bitmapBlocks, bitmaps, reserved, fileSystemBlockSize);
             
-
             // calculate partition start offset
-            var partitionStartByteOffset = lowCyl * surfaces * blocksPerTrack * fileSystemBlockSize;
+            var partitionStartByteOffset = (long)lowCyl * surfaces * blocksPerTrack * fileSystemBlockSize;
 
             // write dos type at partition start
             stream.Seek(partitionStartByteOffset, SeekOrigin.Begin);
